src/software_engineering.py
```python
import os
import logging
from src.cyber_security import applyCyberSecurityMeasures
from src.encryption import encryptData
from src.authentication import authenticateUser

def applySoftwareEngineeringBestPractices(dataSet, userInput, encryptedData, authenticatedUser):
    try:
        # Apply cybersecurity measures
        applyCyberSecurityMeasures()

        # Encrypt data
        encryptedData = encryptData(dataSet)

        # Authenticate user
        authenticatedUser = authenticateUser(userInput)

        # Log successful operations
        logging.info("Data encryption and user authentication successful.")

    except Exception as e:
        # Log any exceptions
        logging.error(f"An error occurred: {str(e)}")

    finally:
        # Ensure all resources are released
        if dataSet is not None:
            dataSet.close()
        if userInput is not None:
            userInput.close()

    return encryptedData, authenticatedUser
```