src/tests/test_software_engineering.py
import unittest
from src import software_engineering

class TestSoftwareEngineering(unittest.TestCase):

    def setUp(self):
        self.dataSet = {"data": "sample data"}
        self.userInput = {"input": "sample input"}

    def test_applySoftwareEngineeringBestPractices(self):
        result = software_engineering.applySoftwareEngineeringBestPractices(self.dataSet, self.userInput)
        self.assertIsNotNone(result)
        self.assertIsInstance(result, dict)

    def test_processData(self):
        result = software_engineering.processData(self.dataSet)
        self.assertIsNotNone(result)
        self.assertIsInstance(result, dict)

    def test_optimizeForMobile(self):
        result = software_engineering.optimizeForMobile(self.dataSet, self.userInput)
        self.assertIsNotNone(result)
        self.assertIsInstance(result, dict)

if __name__ == '__main__':
    unittest.main()